extract d3 specific elements
